Trying to figure out contactMaterial
I don't won't my sprite to bounce off walls, trying to figure out a way
to modify the contactMaterial between the sprite and walls.
https://phaser.io/examples/v2/p2-physics/contact-material
this is not very helpful, but is worth looking at

My problem now is since the walls are generated via 'convertTileMap',
they become an array of body objects, and I don't know how to define
contact material for multiple bodys.